+++ src/examples/src/org/apache/poi/hssf/usermodel/examples/FrillsAndFills.java
@@ -1,4 +1,3 @@
-
 /* ====================================================================
    Licensed to the Apache Software Foundation (ASF) under one or more
    contributor license agreements.  See the NOTICE file distributed with
@@ -15,7 +14,6 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 ==================================================================== */
-        
 
 package org.apache.poi.hssf.usermodel.examples;
 
@@ -30,22 +28,19 @@ import java.io.IOException;
  *
  * @author Glen Stampoultzis (glens at apache.org)
  */
-public class FrillsAndFills
-{
-    public static void main(String[] args)
-            throws IOException
-    {
+public class FrillsAndFills {
+    public static void main(String[] args) throws IOException {
         HSSFWorkbook wb = new HSSFWorkbook();
         HSSFSheet sheet = wb.createSheet("new sheet");
 
         // Create a row and put some cells in it. Rows are 0 based.
-        HSSFRow row = sheet.createRow((short) 1);
+        HSSFRow row = sheet.createRow(1);
 
         // Aqua background
         HSSFCellStyle style = wb.createCellStyle();
         style.setFillBackgroundColor(HSSFColor.AQUA.index);
         style.setFillPattern(HSSFCellStyle.BIG_SPOTS);
-        HSSFCell cell = row.createCell((short) 1);
+        HSSFCell cell = row.createCell(1);
         cell.setCellValue("X");
         cell.setCellStyle(style);
 
@@ -53,7 +48,7 @@ public class FrillsAndFills
         style = wb.createCellStyle();
         style.setFillForegroundColor(HSSFColor.ORANGE.index);
         style.setFillPattern(HSSFCellStyle.SOLID_FOREGROUND);
-        cell = row.createCell((short) 2);
+        cell = row.createCell(2);
         cell.setCellValue("X");
         cell.setCellStyle(style);
