+++ src/examples/src/org/apache/poi/hssf/usermodel/examples/OfficeDrawingWithGraphics.java
@@ -1,4 +1,3 @@
-
 /* ====================================================================
    Licensed to the Apache Software Foundation (ASF) under one or more
    contributor license agreements.  See the NOTICE file distributed with
@@ -15,7 +14,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 ==================================================================== */
-        
+
 package org.apache.poi.hssf.usermodel.examples;
 
 import org.apache.poi.hssf.usermodel.*;
@@ -29,15 +28,13 @@ import java.io.IOException;
  *
  * @author Glen Stampoultzis (glens at apache.org)
  */
-public class OfficeDrawingWithGraphics
-{
-    public static void main( String[] args ) throws IOException
-    {
+public class OfficeDrawingWithGraphics {
+    public static void main( String[] args ) throws IOException {
         // Create a workbook with one sheet and size the first three somewhat
         // larger so we can fit the chemical structure diagram in.
         HSSFWorkbook wb = new HSSFWorkbook();
         HSSFSheet sheet = wb.createSheet( "my drawing" );
-        sheet.setColumnWidth((short)1, (short)(256 * 27));
+        sheet.setColumnWidth(1, 256 * 27);
         HSSFRow row1 = sheet.createRow(0);
         row1.setHeightInPoints(10 * 15);
         HSSFRow row2 = sheet.createRow(1);
@@ -47,9 +44,9 @@ public class OfficeDrawingWithGraphics
 
         // Add some cells so we can test that the anchoring works when we
         // sort them.
-        row1.createCell((short)0).setCellValue("C");
-        row2.createCell((short)0).setCellValue("A");
-        row3.createCell((short)0).setCellValue("B");
+        row1.createCell(0).setCellValue("C");
+        row2.createCell(0).setCellValue("A");
+        row3.createCell(0).setCellValue("B");
 
         // Create the top level drawing patriarch.
         HSSFPatriarch patriarch = sheet.createDrawingPatriarch();
@@ -62,7 +59,7 @@ public class OfficeDrawingWithGraphics
         a = new HSSFClientAnchor( 0, 0, 1023, 255, (short) 1, 0, (short) 1, 0 );
         group = patriarch.createGroup( a );
         group.setCoordinates( 0, 0, 320, 276 );
-        float verticalPointsPerPixel = a.getAnchorHeightInPoints(sheet) / (float)Math.abs(group.getY2() - group.getY1());
+        float verticalPointsPerPixel = a.getAnchorHeightInPoints(sheet) / Math.abs(group.getY2() - group.getY1());
         g = new EscherGraphics( group, wb, Color.black, verticalPointsPerPixel );
         g2d = new EscherGraphics2d( g );
         drawStar( g2d );
@@ -70,7 +67,7 @@ public class OfficeDrawingWithGraphics
         a = new HSSFClientAnchor( 0, 0, 1023, 255, (short) 1, 1, (short) 1, 1 );
         group = patriarch.createGroup( a );
         group.setCoordinates( 0, 0, 640, 276 );
-        verticalPointsPerPixel = a.getAnchorHeightInPoints(sheet) / (float)Math.abs(group.getY2() - group.getY1());
+        verticalPointsPerPixel = a.getAnchorHeightInPoints(sheet) / Math.abs(group.getY2() - group.getY1());
 //        verticalPixelsPerPoint = (float)Math.abs(group.getY2() - group.getY1()) / a.getAnchorHeightInPoints(sheet);
         g = new EscherGraphics( group, wb, Color.black, verticalPointsPerPixel );
         g2d = new EscherGraphics2d( g );
